+++ src/models/job_product.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey, DECIMAL
 from sqlalchemy.orm import relationship
 from src.models.base import Base
 
@@ -12,7 +12,7 @@ class JobProduct(Base):
     quantity = Column(Integer, nullable=False, default=1)
     # Guardamos el precio del producto al momento exacto de crear el JOB. 
     # Asi si en el futuro se edita el producto, las facturas viejas no se arruinan!
-    historic_price = Column(Integer, nullable=False)
+    historic_price = Column(DECIMAL(10, 2), nullable=False)
 
     # Relaciones Back
     job = relationship("Job", back_populates="products")
