+++ src/models/payment.py
@@ -1,5 +1,5 @@
 import enum
-from sqlalchemy import Column, Integer, Boolean, DateTime, ForeignKey, Enum as SqlaEnum, func
+from sqlalchemy import Column, Integer, Boolean, DateTime, ForeignKey, Enum as SqlaEnum, func, DECIMAL
 import datetime
 from sqlalchemy.orm import relationship
 from src.models.base import Base
@@ -17,7 +17,7 @@ class Payment(Base):
     id_invoice = Column(Integer, ForeignKey("invoices.id_invoice"), nullable=False)
     
     pay_date = Column(DateTime, nullable=False, default=datetime.datetime.now) # Actualizado de Date a DateTime
-    payment_amount = Column(Integer, nullable=False)
+    payment_amount = Column(DECIMAL(10, 2), nullable=False)
     payment_type = Column(SqlaEnum(PaymentType), nullable=False, default=PaymentType.EFECTIVO)
     status = Column(Boolean, nullable=False, default=True)
