+++ src/models/product.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean
+from sqlalchemy import Column, Integer, String, DECIMAL, Boolean
 from src.models.base import Base
 
 class Product(Base):
@@ -6,7 +6,7 @@ class Product(Base):
 
     id_product = Column(Integer, primary_key=True, index=True, autoincrement=True)
     name = Column(String(50), nullable=False, unique=True)
-    price = Column(Integer, nullable=False)
+    price = Column(DECIMAL(10, 2), nullable=False)
     
     # Representamos production time en días (integer)
     production_time = Column(Integer, nullable=False)
